Move all content from NewHomePage into Home.tsx
